\name{F1}
\alias{F1}
\title{
Labeling of memberships of
two groups
}
\description{
Internal function to label the membership of two clusters
}
\usage{
F1(A, B)
}
\arguments{
  \item{A}{
Memberships in the first cluster
}
  \item{B}{
Membership in the second cluster
}
}
\value{
Produced a two-column matrix populated with zeros and ones to indicate what IDs (row names) are included in each of two vectors in which some element may be shared
}

\author{
Roger Guevara (roger.guevara@inecol.mx)
}

